+++ src/client/components/SelectWithLabel.tsx
@@ -1,11 +1,13 @@
+import { memo } from "react";
+
 import {
   Select,
   SelectTrigger,
   SelectValue,
   SelectContent,
   SelectItem,
-} from "./ui/select";
-import { Label } from "./ui/label";
+  Label,
+} from "./ui";
 
 interface SelectWithLabelProps {
   label: string;
@@ -15,31 +17,25 @@ interface SelectWithLabelProps {
   onChange: (value: string) => void;
 }
 
-const SelectWithLabel = ({
-  label,
-  value,
-  items,
-  placeholder,
-  onChange,
-}: SelectWithLabelProps) => {
-  return (
-    <div className="grid w-full max-w-sm items-center gap-1.5">
-      <Label>{label}</Label>
-      <Select value={value} onValueChange={onChange}>
-        <SelectTrigger>
-          <SelectValue placeholder={placeholder} />
-        </SelectTrigger>
+export const SelectWithLabel = memo(
+  ({ label, value, items, placeholder, onChange }: SelectWithLabelProps) => {
+    return (
+      <div className="grid w-full max-w-sm items-center gap-1.5">
+        <Label>{label}</Label>
+        <Select value={value} onValueChange={onChange}>
+          <SelectTrigger>
+            <SelectValue placeholder={placeholder} />
+          </SelectTrigger>
 
-        <SelectContent>
-          {items.map((item) => (
-            <SelectItem key={item} value={item}>
-              {item}
-            </SelectItem>
-          ))}
-        </SelectContent>
-      </Select>
-    </div>
-  );
-};
-
-export default SelectWithLabel;
+          <SelectContent>
+            {items.map((item) => (
+              <SelectItem key={item} value={item}>
+                {item}
+              </SelectItem>
+            ))}
+          </SelectContent>
+        </Select>
+      </div>
+    );
+  }
+);
