+++ src/client/components/index.ts
@@ -1,0 +1,3 @@
+export { EditableCell } from "./EditableCell";
+export { InputWithLabel } from "./InputWithLabel";
+export { SelectWithLabel } from "./SelectWithLabel";
