+++ src/client/components/ui/index.ts
@@ -1,0 +1,12 @@
+export { Button } from "./button";
+export { Label } from "./label";
+export { Input } from "./input";
+export {
+  Select,
+  SelectContent,
+  SelectTrigger,
+  SelectItem,
+  SelectValue,
+} from "./select";
+export { Table, TableBody, TableCell, TableRow } from "./table";
+export { Toaster } from "./toaster";
